refactor(ipc): adapted IPC work to new project structure
Adapted IPC work to project layout defined in #126

#27
