+++ app/main/api/handlers/ping.ts
@@ -1,4 +1,4 @@
-import * as IPCMethods from "../../ipc/common/methods"
+import * as IPCMethods from "../../../../app/common/methods"
 
 /**
  * Handler of ping method
