+++ config/webpack.main.js
@@ -14,7 +14,7 @@ const baseConfig = {
   devtool: "source-map",
   target: "electron-main",
 
-  entry: [path.resolve(__dirname, "../app/main/index.js")],
+  entry: [path.resolve(__dirname, "../app/main/index.ts")],
 
   resolve: {
     extensions: [".js", ".ts", ".json"]
