+++ test/app/common/methods.spec.ts
@@ -1,4 +1,4 @@
-import * as IPCMethods from "app/lib/api/ipc/common/methods"
+import * as IPCMethods from "appCommon/methods"
 import * as fixtures from "./methodsFixtures"
 
 describe(`IPCMethods lib: validate valid method descriptors" +
